cdk/requirements.txt
-e .
aws_cdk.aws_codecommit
aws_cdk.aws_cloudfront
aws_cdk.aws_iam
aws_cdk.aws_s3
aws_cdk.aws_s3_deployment
aws_cdk.aws_ecr
